chore: format shell script

+++ src/burn-notice.sh
@@ -26,12 +26,12 @@ BURN NOTICE COMPLETE! MAKE SURE TO CLOSE THIS TERMINAL!
 \n\nTo fully burn your identity, it's suggested to also remove browsing data, log out of iCloud and other installed apps, and if very paranoid, format your hard drive. Remember, this script did not remove your trash, browser data, or log you out of iCloud or other accounts."
 }
 
-start_countdown() {    
+start_countdown() {
     # Grace period allowing a user to quit if they accidentally continued
     local seconds
     local second
     seconds=(5 4 3 2 1)
-    for second in "${seconds[@]}" ; do
+    for second in "${seconds[@]}"; do
         echo "Burning in $second seconds..."
         sleep 1
     done
@@ -41,8 +41,8 @@ start_countdown() {
 remove_hidden_items() {
     # Remove hidden files/folders
     local item
-    for item in .[^.]* ; do
-        if [[ -d "$item" ]] ; then
+    for item in .[^.]*; do
+        if [[ -d "$item" ]]; then
             cd "$item" || exit 1
             rm -rf ./*
             rm -rf .[^.]*
@@ -58,23 +58,23 @@ remove_hidden_items() {
 remove_visible_items() {
     # Remove visible files/folders
     local item
-    for item in * ; do
+    for item in *; do
         # Ignore some of the most common cloud storage provider folders
-        if ! [[ \
-            "$item" = "Backup and Sync" || \
-            "$item" = "Box" || \
-            "$item" = "dropbox" || \
-            "$item" = "Dropbox" || \
-            "$item" = "Google Drive File Stream" || \
-            "$item" = "Google Drive" || \
-            "$item" = "My Drive" || \
-            "$item" = "nextcloud" || \
-            "$item" = "Nextcloud" || \
-            "$item" = "OneDrive" || \
-            "$item" = "owncloud" || \
-            "$item" = "Owncloud" \
-        ]] ; then
-            if [[ -d "$item" ]] ; then
+        if ! [[ 
+            "$item" = "Backup and Sync" ||
+            "$item" = "Box" ||
+            "$item" = "dropbox" ||
+            "$item" = "Dropbox" ||
+            "$item" = "Google Drive File Stream" ||
+            "$item" = "Google Drive" ||
+            "$item" = "My Drive" ||
+            "$item" = "nextcloud" ||
+            "$item" = "Nextcloud" ||
+            "$item" = "OneDrive" ||
+            "$item" = "owncloud" ||
+            "$item" = "Owncloud" ]] \
+            ; then
+            if [[ -d "$item" ]]; then
                 cd "$item" || exit 1
                 rm -rf ./*
                 rm -rf .[^.]*
